Добавлены интерфейсные классы, чтобы Renderer для WMF и EMF был общим.
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@62301 954022d7-b5bf-4e40-9824-e11837661b57

+++ DesktopEditor/raster/Metafile/Wmf/WmfPlayer.cpp
@@ -121,7 +121,7 @@ namespace MetaFile
 				if (ushIndex > m_vAvailableIndexes[nIterator])
 				{
 					ushIndex = m_vAvailableIndexes[nIterator];
-					nAvailableIndex = nIterator;
+					nAvailableIndex = (int)nIterator;
 				}
 			}
 		}
